Added empty favorites component. Split login component.
